LPS-142497 Extract TopperLabel
